requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
requests==2.31.0
pytest==8.0.0